Block Amazonbot due to excessive crawling
Heavy crawling raising load on servers (and using
up AWS CPU credits) far beyond the crawl delay.

Understaing that most crawlers don't follow the delay
from the robots file, the level of calls was far
in excess of acceptable use. Not as bad as GoogleOther
bot, but enough to prevent CPU credits from recovering.

Props to bingbot for being an okay bot.